refactor(reporting): move the AI cluster to views/reporting/ai.py
Move the 4 /api/reporting/ai/* routes (ask/build/agent/caption) and their
private helpers (_ai_config, _ai_schema_text, _ai_catalog_text,
_validate_definition_for_user, _normalize_definition, _audit_ai,
_ai_daily_limit, _ai_asks_today, plus _extract_agent_artifacts and
_caption_columns) out of nx_lib/views/reporting/__init__.py into a new
nx_lib/views/reporting/ai.py, per beautify-phase-2a Task 1.

- ai.py imports its own has_permission/get_locale/etc.; anything it still
  needs from the package (_get_effective_source, _accessible_sql_targets,
  _metrics_for_source, _allowed_processes, _load_process_configs,
  _load_field_col_maps, _accessible_curated_sources, _has_acked,
  _authorize_sql_target, _audit_sql, _run_sql, _SQL_TARGETS) is fetched via
  a local 'from . import ...' inside the function that needs it, to avoid
  a circular import at module load (same trick nx_lib/reporting/runner.py
  already uses for the same class of problem).
- __init__.py re-exports every moved name (via 'from .ai import ...' plus
  an __all__ list, matching the nx_lib/views/admin package precedent) so
  nx_lib.views.reporting.<name> keeps resolving for callers, tests, and
  nx_lib/reporting/runner.py's other attribute reads. register_routes()
  now delegates the 4 AI routes to ai.register_routes(app) in the same
  relative position.
- Fixed two latent __file__-relative-depth bugs the package move exposed:
  _GUIDE_MD's parents[2] (now REPO_ROOT-based, depth-independent) and the
  bare 'from ..X' imports that needed to become 'from ...X' one level down.
- Retargeted every nx_lib.views.reporting.<moved-name> monkeypatch in
  tests/integration/test_reporting_ai_routes.py (including has_permission,
  ai_ask/ai_ask_definition/ai_caption, ask_agentic/ask_agentic_iter,
  make_agent_step, get_locale, fetch_docprocessing_catalog — all now bound
  in ai.py's own module namespace) to nx_lib.views.reporting.ai.<name>.
  Verified live via a vacuous-pass probe: pointing one retargeted patch
  back at the old (now-wrong) path made its test fail immediately.

tests/integration/test_reporting_ai_routes.py, test_reporting_routes.py,
test_reporting_metrics_routes.py, and the reporting unit tests are green
(187 passed; one pre-existing unrelated flake in
test_ai_agent_empty_answer_gets_artifact_aware_fallback reproduces
identically on the pre-Task-1 commit).
